db: fix checkpoint set / list
Need to use dateTimeToJson / dateTimeFromJson to interact with the
checkpoint value in the DB.

Fixes issue #9.
